chore: Restructure Application runtime.
